% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_column_codes_from_score_file.R
\name{extract_column_codes_from_score_file}
\alias{extract_column_codes_from_score_file}
\title{Extract column codes from a csv score file}
\usage{
extract_column_codes_from_score_file(csv_filepath)
}
\arguments{
\item{csv_filepath}{the path of the csv file}
}
\value{
column codes vector with NA for the column without codes
}
\description{
Extract column codes from a csv score file
}
\examples{
extract_column_codes_from_score_file("score.csv", "ssr", "oqn")
}
