% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table.R
\name{format_table}
\alias{format_table}
\title{Format table}
\usage{
format_table(
  table,
  finess,
  selected_columns,
  translated_columns,
  filters,
  row_names,
  rows_translated,
  proper_left_col
)
}
\description{
Format table
}
