% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_volume_table.R
\name{format_volume_table}
\alias{format_volume_table}
\title{Format volume table}
\usage{
format_volume_table(champ = "mco", statut = "dgf", finess)
}
\arguments{
\item{statut}{}
}
\description{
Format volume table
}
