% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ovalide_tables.R
\name{load_ovalide_tables}
\alias{load_ovalide_tables}
\title{Load ovalide table stored data in `the` internal state}
\usage{
load_ovalide_tables(nature, force = FALSE)
}
\value{
list of ovalide table dataframes
}
\description{
Load ovalide table stored data in `the` internal state
}
