% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ovalide_tables.R
\name{read_ovalide_tables}
\alias{read_ovalide_tables}
\title{Read ovalide table stored data}
\usage{
read_ovalide_tables(champ = "mco", statut = "dgf")
}
\arguments{
\item{statut}{}
}
\value{
list of ovalide table dataframes
}
\description{
Read ovalide table stored data
}
