% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_score_csv_file.R
\name{read_score_csv_file}
\alias{read_score_csv_file}
\title{Read a csv score file and affect to the_environment}
\usage{
read_score_csv_file(
  csv_filepath,
  champ = c("mco", "had", "ssr", "psy"),
  statut = c("dgf", "oqn")
)
}
\arguments{
\item{csv_filepath}{the path of the csv file}

\item{champ}{name of the PMSI champ}

\item{statut}{private or public status}
}
\description{
Read a csv score file and affect to the_environment
}
\examples{
read_score_file("score.csv", "psy", "dgf")
}
