% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zip_table_file.R
\name{read_zip_table_file}
\alias{read_zip_table_file}
\title{Read table data from zip ovalide file}
\usage{
read_zip_table_file(zip_filepath, nature)
}
\arguments{
\item{zip_filepath}{the zip file path}

\item{champ}{the PMSI champ}

\item{statut}{private or public status}
}
\description{
Read table data from zip ovalide file
}
\examples{
read_zip_table_file("file.zip", "had", "oqn")

# can be called with `progressr::with_progress`
with_progressr(read_zip_table_file("file.zip", "had", "oqn"))

# can also be called with `progressr::withProgressShiny`
}
